Flytt util-funksjonalitet ut av typefil og inn i util-fil
