+++ src/frontend/Sider/Behandling/Inngangsvilkår/typer/vilkårperiode/aktivitet.ts
@@ -8,8 +8,6 @@ import {
     AktivitetLæremidlerFaktaOgSvar,
     AktivitetLæremidlerFaktaOgVurderinger,
 } from './aktivitetLæremidler';
-import { SelectOption } from '../../../../../komponenter/Skjema/SelectMedOptions';
-import { Stønadstype } from '../../../../../typer/behandling/behandlingTema';
 
 export type Aktivitet = AktivitetBarnetilsyn | AktivitetLæremidler;
 
@@ -25,41 +23,6 @@ export const AktivitetTypeTilTekst: Record<AktivitetType, string> = {
     UTDANNING: 'Utdanning',
     REELL_ARBEIDSSØKER: 'Reell arbeidssøker',
     INGEN_AKTIVITET: 'Ingen aktivitet',
-};
-
-export const aktivitetTypeTilTekst = (type: AktivitetType | '') => {
-    if (type === '') return type;
-
-    return AktivitetTypeTilTekst[type];
-};
-
-export const lagAktivitetTypeOptions = (stønadstype: Stønadstype): SelectOption[] => {
-    const relevanteTyper = finnRelevanteAktivitetTyperForStønad(stønadstype);
-
-    return relevanteTyper.map((type) => ({
-        value: type,
-        label: AktivitetTypeTilTekst[type],
-    }));
-};
-
-export const lagAktivitetTypeOptionsForStønadsperiode = (stønadstype: Stønadstype) =>
-    lagAktivitetTypeOptions(stønadstype).filter(
-        (option) => option.value !== AktivitetType.INGEN_AKTIVITET
-    );
-
-export const finnRelevanteAktivitetTyperForStønad = (stønadstype: Stønadstype): AktivitetType[] => {
-    switch (stønadstype) {
-        case Stønadstype.BARNETILSYN:
-            return [
-                AktivitetType.TILTAK,
-                AktivitetType.UTDANNING,
-                AktivitetType.REELL_ARBEIDSSØKER,
-                AktivitetType.INGEN_AKTIVITET,
-            ];
-
-        case Stønadstype.LÆREMIDLER:
-            return [AktivitetType.TILTAK, AktivitetType.UTDANNING, AktivitetType.INGEN_AKTIVITET];
-    }
 };
 
 export type AktivitetFaktaOgVurderinger =
