+++ src/frontend/Sider/Behandling/VedtakOgBeregning/Barnetilsyn/InnvilgeVedtak/Beregningsresultat.tsx
@@ -8,7 +8,7 @@ import '@navikt/ds-css';
 
 import { BeregningsresultatTilsynBarn, Vedtaksperiode } from '../../../../../typer/vedtak';
 import { formaterIsoPeriode } from '../../../../../utils/dato';
-import { aktivitetTypeTilTekst } from '../../../Inngangsvilkår/typer/vilkårperiode/aktivitet';
+import { aktivitetTypeTilTekst } from '../../../Inngangsvilkår/Aktivitet/utilsAktivitet';
 import { målgruppeTypeTilTekst } from '../../../Inngangsvilkår/typer/vilkårperiode/målgruppe';
 
 const Container = styled.div`
